fix: update merge
